requirements.txt
cycler==0.10.0
Cython==0.29.21
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.1
opencv-python==4.4.0.42
pandas==1.1.0
Pillow==7.2.0
protobuf==3.12.2
pycocotools==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scipy==1.5.2
six==1.15.0
tensorboardX==2.1
torch==1.5.1
torchvision==0.6.1
tqdm==4.48.0
